+++ Profile/Shopware6/Converter/CategoryProductStreamAssociationConverter.php
@@ -1,0 +1,38 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
+
+use Shopware\Core\Framework\Log\Package;
+use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\MigrationContextInterface;
+use SwagMigrationAssistant\Profile\Shopware6\DataSelection\DataSet\CategoryProductStreamAssociationDataSet;
+use SwagMigrationAssistant\Profile\Shopware6\Shopware6MajorProfile;
+
+#[Package('services-settings')]
+class CategoryProductStreamAssociationConverter extends ShopwareConverter
+{
+    public function supports(MigrationContextInterface $migrationContext): bool
+    {
+        return $migrationContext->getProfile()->getName() === Shopware6MajorProfile::PROFILE_NAME
+            && $this->getDataSetEntity($migrationContext) === CategoryProductStreamAssociationDataSet::getEntity();
+    }
+
+    protected function convertData(array $data): ConvertStruct
+    {
+        $converted = $data;
+
+        $this->mainMapping = $this->getOrCreateMappingMainCompleteFacade(
+            DefaultEntities::CATEGORY,
+            $data['id'],
+            $converted['id']
+        );
+
+        return new ConvertStruct($converted, null, $this->mainMapping['id']);
+    }
+}