+++ Profile/Shopware6/Converter/SeoUrlConverter.php
@@ -10,9 +10,12 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\MigrationContextInterface;
+use SwagMigrationAssistant\Profile\Shopware6\DataSelection\DataSet\SeoUrlDataSet;
+use SwagMigrationAssistant\Profile\Shopware6\Shopware6MajorProfile;
 
 #[Package('services-settings')]
-abstract class SeoUrlConverter extends ShopwareConverter
+class SeoUrlConverter extends ShopwareConverter
 {
     /**
      * @var string
@@ -23,6 +26,12 @@ abstract class SeoUrlConverter extends ShopwareConverter
      * @var string
      */
     protected const CATEGORY_ROUTE_NAME = 'frontend.navigation.page';
+
+    public function supports(MigrationContextInterface $migrationContext): bool
+    {
+        return $migrationContext->getProfile()->getName() === Shopware6MajorProfile::PROFILE_NAME
+            && $this->getDataSetEntity($migrationContext) === SeoUrlDataSet::getEntity();
+    }
 
     protected function convertData(array $data): ConvertStruct
     {