+++ Profile/Shopware6/DataSelection/DataSet/CategoryProductStreamAssociationDataSet.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SwagMigrationAssistant\Profile\Shopware6\DataSelection\DataSet;
+
+use Shopware\Core\Framework\Log\Package;
+use SwagMigrationAssistant\Migration\DataSelection\DataSet\DataSet;
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\MigrationContextInterface;
+use SwagMigrationAssistant\Profile\Shopware6\Shopware6ProfileInterface;
+
+#[Package('services-settings')]
+class CategoryProductStreamAssociationDataSet extends DataSet
+{
+    public static function getEntity(): string
+    {
+        return DefaultEntities::CATEGORY_PRODUCT_STREAM_ASSOCIATION;
+    }
+
+    public function supports(MigrationContextInterface $migrationContext): bool
+    {
+        return $migrationContext->getProfile() instanceof Shopware6ProfileInterface;
+    }
+}