+++ Profile/Shopware6/Gateway/Api/Reader/CategoryProductStreamAssociationReader.php
@@ -1,0 +1,20 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SwagMigrationAssistant\Profile\Shopware6\Gateway\Api\Reader;
+
+use Shopware\Core\Framework\Log\Package;
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+
+#[Package('services-settings')]
+class CategoryProductStreamAssociationReader extends ApiReader
+{
+    protected function getIdentifier(): string
+    {
+        return DefaultEntities::CATEGORY_PRODUCT_STREAM_ASSOCIATION;
+    }
+}