+++ Test/_fixtures/Shopware6/ShippingMethod/02-ExistingCase/mapping.php
@@ -1,0 +1,36 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+
+return [
+    [
+        'entityName' => DefaultEntities::LANGUAGE,
+        'oldIdentifier' => 'a3632dd1442e441eb0a98adb86885ae0',
+        'newIdentifier' => '41248655c60c4dd58cde43f14cd4f149',
+    ],
+    [
+        'entityName' => DefaultEntities::LANGUAGE,
+        'oldIdentifier' => '2fbb5fe2e29a4d70aa5854ce7ce3e20b',
+        'newIdentifier' => '5dd637353d044752ae6a8c6e7f53430b',
+    ],
+    [
+        'entityName' => DefaultEntities::TAX,
+        'oldIdentifier' => 'f5e4c4d734314892bf37acd3ebae9eee',
+        'newIdentifier' => 'de0fd756f9314611a2444bd39da75287',
+    ],
+    [
+        'entityName' => DefaultEntities::CURRENCY,
+        'oldIdentifier' => 'b7d2554b0ce847cd82f3ac9bd1c0dfca',
+        'newIdentifier' => 'b8d2554b0ce847cd82f3ac9bd1c0dfca',
+    ],
+    [
+        'entityName' => DefaultEntities::DELIVERY_TIME,
+        'oldIdentifier' => '80293267cbeb4513bc4d3cede0144022',
+        'newIdentifier' => 'bdeb106f47ab4255b2bd5f35d84cae7c',
+    ],
+];